feat: clean name
