--- RenewXControl.Console/Configuration/AssetsModel/Assets/WindTurbineConfig.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace RenewXControl.Console.InitConfiguration.AssetsModelConfig.Assets
-{
-    public class WindTurbineConfig    
-    {
-        public double SetPoint { get; set; }
-        public double WindSpeed { get; set; }
-        public double ActivePower { get; set; }
-    }
-}
