--- RenewXControl.Console/Configuration/AssetsModel/Users/SiteConfig.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using RenewXControl.Console.InitConfiguration.AssetsModelConfig.Assets;
-
-namespace RenewXControl.Console.InitConfiguration.AssetsModelConfig.Users
-{
-    public class SiteConfig
-    {
-        public string Name { get; set; }
-        public string Location { get; set; }
-    }
-}
