--- RenewXControl.Console/Configuration/AssetsModel/Users/UserConfig.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using RenewXControl.Console.Domain.Users;
-
-namespace RenewXControl.Console.InitConfiguration.AssetsModelConfig.Users
-{
-    public class UserConfig
-    {
-        public string Name { get; set; }
-    }
-}
