--- RenewXControl.Console/RenewXControl.Console.csproj
@@ -1,13 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-<!--Please consider adding namespace and dll names here. Also, the library name (RenewXControl.Console) has an unnecessary project name (RenewXControl) in it.
-Please remove it while you have that name as your solution name but keep it in the namespace and dll name.
-So, your namespace and dll name can be like this: RenewXControl.Console but the library name should be Console.
-We will talk more about this while talking about the architecture-->
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net8.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-</Project>
