+++ RenewXControl.sln
@@ -2,7 +2,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.13.35919.96
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RenewXControl.Console", "RenewXControl.Console\RenewXControl.Console.csproj", "{31280121-243E-4EE0-A4EF-21A500DFDAAB}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RenewXControl", "RenewXControl\RenewXControl.csproj", "{31280121-243E-4EE0-A4EF-21A500DFDAAB}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
