+++ RenewXControl/Configuration/AssetsModel/Assets/BatteryConfig.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace RenewXControl.Console.InitConfiguration.AssetsModelConfig.Assets
+﻿namespace RenewXControl.Configuration.AssetsModel.Assets
 {
     // A config object is not supposed to have any behavior. It only carries some read-only data
     // So, it is better to define it as a record class. Please investigate about record, struct and class in C#
