+++ RenewXControl/Configuration/AssetsModel/Assets/SolarPanelConfig.cs
@@ -1,0 +1,9 @@
+﻿namespace RenewXControl.Configuration.AssetsModel.Assets
+{
+    public class SolarPanelConfig
+    {
+        public double Irradiance { get; set; }
+        public double SetPoint { get; set; }
+        public double ActivePower { get; set; }
+    }
+}
