+++ RenewXControl/Configuration/AssetsModel/Assets/WindTurbineConfig.cs
@@ -1,0 +1,9 @@
+﻿namespace RenewXControl.Configuration.AssetsModel.Assets
+{
+    public class WindTurbineConfig    
+    {
+        public double SetPoint { get; set; }
+        public double WindSpeed { get; set; }
+        public double ActivePower { get; set; }
+    }
+}
