+++ RenewXControl/Configuration/AssetsModel/Users/SiteConfig.cs
@@ -1,0 +1,8 @@
+﻿namespace RenewXControl.Configuration.AssetsModel.Users
+{
+    public class SiteConfig
+    {
+        public string Name { get; set; }
+        public string Location { get; set; }
+    }
+}
