+++ RenewXControl/Configuration/AssetsModel/Users/UserConfig.cs
@@ -1,0 +1,7 @@
+﻿namespace RenewXControl.Configuration.AssetsModel.Users
+{
+    public class UserConfig
+    {
+        public string Name { get; set; }
+    }
+}
