+++ RenewXControl/Configuration/JsonFiles/SolarPanel.json
