+++ RenewXControl/doc/Assets/Solar Panel Workflow.md
