+++ RenewXControl/doc/Configuration.md
